2-file scheme adopted code decoupled - CSS has to be moved to theme file
